Fix ethNetworkToggle when scrolled down on page
This was a hard bug to track down;
ethNetworkToggle works fine when clicked at the top of the page,
but it hides the entire header when the page is even slightly scrolled down!
This is apparently caused by 'overflow-x: hidden' on the html body,
so we removed that in order to fix the bug.