Fix LoadingIndicator layout shift
Hard to find the cause:
The layout shift is caused because the <ul> is a block element by default.
Apparently this causes the ul to render below the span initially,
before the CSS is loaded, and only afterwards correcting for flex-row,
and moving sideways of the span. CSS fixes did not work at all.
The only possible solution is to use an inline-default element like <span>!