Fix TokenDropdown bug where MenuItems doesn't grow when content requires it;
The MenuItems Popover is always the same width as the MenuButton,
which can be too small for long token symbols with large token amounts.
It should really be expanding, to allow for more room for these specific
tokens. This bug was caused because of the 'break-all' set in the
PageWrapper component, which causes all text to wrap and so breaks the
browser's ability to correctly size the MenuItems!