Identify problem why ErrorStates are returning LoadingIndicators;
The change that modified the return type of DataStateBase.Error from ErrorStateBase
to DataStateBase (line 46 in data-state.tsx) causes Typescript to mistype the
ErrorStateBases as ValueStateBases, so we end up with a ValueStateBase with an
Error and undefined value. This in turn causes the Render method to simply render
a LoadingIndicator, because it sees a ValueStateBase with an undefined value!
We need some way for Typescript to be able to correctly identify the ErrorStateBase as
ErrorStateBase instead of a ValueStateBase ...